Improve mcu upscaling

+++ src/org/terifan/imageio/jpeg/test/MeasureErrorRate.java
@@ -1,19 +1,13 @@
 package org.terifan.imageio.jpeg.test;
 
 import java.awt.image.BufferedImage;
-import java.io.ByteArrayInputStream;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import javax.imageio.ImageIO;
-import org.terifan.imageio.jpeg.decoder.JPEGImageReaderImpl;
 
 
 public class MeasureErrorRate
 {
 	static double measureError(BufferedImage aImage, BufferedImage aComparisonImage)
 	{
-		int accumError = 0;
+		long accumError = 0;
 
 		for (int y = 0; y < aImage.getHeight(); y++)
 		{
@@ -21,15 +15,19 @@ public class MeasureErrorRate
 			{
 				int c0 = aImage.getRGB(x, y);
 				int c1 = aComparisonImage.getRGB(x, y);
+
 				int r0 = 255 & (c0 >> 16);
 				int g0 = 255 & (c0 >> 8);
 				int b0 = 255 & (c0);
+
 				int r1 = 255 & (c1 >> 16);
 				int g1 = 255 & (c1 >> 8);
 				int b1 = 255 & (c1);
+
 				int er = Math.abs(r0 - r1);
 				int eg = Math.abs(g0 - g1);
 				int eb = Math.abs(b0 - b1);
+
 				accumError += er + eg + eb;
 			}
 		}
